+++ src/Fantomas.Core/CodePrinter.fs
@@ -57,6 +57,9 @@ let rec (|UppercaseExpr|LowercaseExpr|) (expr: Expr) =
     | Expr.DotIndexedGet node -> (|UppercaseExpr|LowercaseExpr|) node.ObjectExpr
     | Expr.TypeApp node -> (|UppercaseExpr|LowercaseExpr|) node.Identifier
     | Expr.Dynamic node -> (|UppercaseExpr|LowercaseExpr|) node.FuncExpr
+    | Expr.AppSingleParenArg node -> (|UppercaseExpr|LowercaseExpr|) node.FunctionExpr
+    | Expr.Paren node -> (|UppercaseExpr|LowercaseExpr|) node.Expr
+    | Expr.App node -> (|UppercaseExpr|LowercaseExpr|) node.FunctionExpr
     | _ -> failwithf "cannot determine if Expr %A is uppercase or lowercase" expr
 
 let (|ParenExpr|_|) (e: Expr) =
@@ -1896,6 +1899,7 @@ let genTupleMultiline (node: ExprTupleNode) =
                 | Expr.Lambda _ -> true
                 | Expr.InfixApp node ->
                     match node.RightHandSide with
+                    | Expr.Match _
                     | Expr.Lambda _ -> true
                     | _ -> false
                 | Expr.SameInfixApps node ->
@@ -2160,7 +2164,7 @@ let genMultilineInfixExpr (node: ExprInfixAppNode) =
         match node.LeftHandSide with
         | IsIfThenElse _ when (ctx.Config.IndentSize - 1 <= node.Operator.Text.Length) ->
             autoParenthesisIfExpressionExceedsPageWidth (genExpr node.LeftHandSide) ctx
-        | Expr.Match _ when (ctx.Config.IndentSize <= node.Operator.Text.Length) ->
+        | Expr.Match _ when (ctx.Config.IndentSize - 1 <= node.Operator.Text.Length) ->
             let ctxAfterMatch = genExpr node.LeftHandSide ctx
 
             let lastClauseIsSingleLine =
@@ -2983,6 +2987,12 @@ let genBinding (b: BindingNode) (ctx: Context) : Context =
 
             let genDestructedTuples =
                 expressionFitsOnRestOfLine (genPat pat) (sepOpenT +> genPat pat +> sepCloseT)
+                +> optSingle
+                    (fun (rt: BindingReturnInfoNode) ->
+                        genSingleTextNode rt.Colon
+                        +> sepSpace
+                        +> atCurrentColumnIndent (genType rt.Type))
+                    b.ReturnType
 
             genXml b.XmlDoc
             +> genAttrAndPref