+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## 6.3.9 - 2024-06-10
+
+### Fixed
+* Invalid location of comma when using named parameters for a class. [#2865](https://github.com/fsprojects/fantomas/issues/2865)
+
+## 6.3.8 - 2024-06-06
+
+### Fixed
+* Fix loss of tuple type annotation without parens. [#2942](https://github.com/fsprojects/fantomas/issues/2942)
+* Fix precedence change of `||>` due to inserted newline. [#2866](https://github.com/fsprojects/fantomas/issues/2866)
+
+## 6.3.7 - 2024-06-01
+
+### Fixed
+* Fix case determination issue with ExprAppNode and ExprParenNode. [#2998](https://github.com/fsprojects/fantomas/issues/2998)
+
+## 6.3.6 - 2024-06-01
+
+### Fixed
+* Fix case determination issue with ExprAppSingleParenArgNode. [#3088](https://github.com/fsprojects/fantomas/issues/3088)
+
 ## 6.3.5 - 2024-05-30
 
 ### Fixed